branch:      main Added in ratings, in a dubious manner. Added a couple of _() calls for i18n
There are issues with duplicate devices, and how to check for them.  I'm kludging that over by hiding them in the 'show' page, and then giving the same rating to each device.  There are also issues with the widget used to select the rating, so there are more kludges in there to fix some bugs with TurboGears, namely making it able to update some trivial properties, such that I don't have to create a new widget every time, but simple redisplay it on a page.  We could really use a better rating widget, perhaps one like the one we have in the GTK gui version.