branch:      main Added some better multi distro support (yet undocumented) with config options for distro-specific things
There is a new option for the location of the hw-uuid file, pretty straight forwarding.  There is now logic code in the config.py file which the packager will have to edit to select the right distro, however upstream will have most of the fun ones commented out.  That code is pretty straight forward.  There is also a string change, pretty minor, but yes, it's a PITA for translation.