branch:      main Added support for more things using SQL alchemy, and converted the add protocol to use json, while leaving in snippets from the old code for backwards compatibility.
add_json, and theoretically add and addDevices work in smoon, as does stats and show.  devices does not work, and will need some revamping to work with a proper DB setup anyways.  the json protocol is simple, and there is probably room to allow for using the older protocol as well in the event that we can't upgrade all clients easily.  we may need a plan to maintain such backwards compatibility anyways.