branch:      main Fuzzled the po files somehow.  Wrote the first set of unit tests.  Switched the model from using assign_mapper to a proper session context.
The unit test is far from complete, but in writing it. I found errors cropped up in what was being commited, and what wasn't to the DB, despite a supposed best practice found in some TG documentation.  It is now fixed, things are flushed and deleted at the appropriate time, and not later.